+++ example/topic_search.dart
@@ -1,0 +1,4 @@
+import 'package:dartpedia/dartpedia.dart' as wiki;
+
+main() async {
+}