+++ CHANGELOG.md
@@ -11,3 +11,7 @@ Added additional documentation to `page.dart`
 
 ## 1.0.3
 Added additional summary functions to `page.dart`
+
+Added Unit Tests for `page.dart`
+
+Updated the /example README to contain code for the examples found in the folder